--- src/golang/lib/models/shared/level.go
@@ -1,11 +1,0 @@
-package shared
-
-type Level string
-
-const (
-	SuccessLevel Level = "success"
-	WarningLevel Level = "warning"
-	ErrorLevel   Level = "error"
-	InfoLevel    Level = "info"
-	NeutralLevel Level = "neutral"
-)
