--- src/golang/lib/models/shared/object.go
@@ -1,9 +1,0 @@
-package shared
-
-type Object string
-
-const (
-	WorkflowObject          Object = "workflow"
-	WorkflowDagResultObject Object = "workflow_dag_result"
-	OrganizationObject      Object = "organization"
-)
