--- src/golang/lib/models/shared/status.go
@@ -1,8 +1,0 @@
-package shared
-
-type Status string
-
-const (
-	UnreadStatus   Status = "unread"
-	ArchivedStatus Status = "archived"
-)
